+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- ✨(mailbox) remove secondary email as required field
+
 ## [1.15.0] - 2025-04-04
 
 ### Added