+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to
 - ✨(domains) define domain check interval as a settings
 - ✨(oidc) add simple introspection backend #832
 - 🧑‍💻(tasks) run management commands #814
+- ✨(mailbox) allow to reset password on mailboxes #834
 
 ### Changed